requirements - Copie.txt
openai-agents
python-dotenv
requests
streamlit
